Add About to Nav
